+++ src/module/sheets/TwodsixActorSheet.ts
@@ -1,9 +1,8 @@
 import { AbstractTwodsixActorSheet } from "./AbstractTwodsixActorSheet";
 import { TWODSIX } from "../config";
-import TwodsixItem from "../entities/TwodsixItem";
+import TwodsixItem, { onRollDamage } from "../entities/TwodsixItem";
 import TwodsixActor from "../entities/TwodsixActor";
-import { DICE_ROLL_MODES } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/constants.mjs";
-import { Consumable, Skills } from "../../types/template";
+import {Consumable, Skills} from "../../types/template";
 import { resolveUnknownAutoMode } from "../utils/rollItemMacro";
 import { getKeyByValue } from "../utils/sheetUtils";
 
@@ -45,7 +44,8 @@ export class TwodsixActorSheet extends AbstractTwodsixActorSheet {
       lifebloodInsteadOfCharacteristics: game.settings.get('twodsix', 'lifebloodInsteadOfCharacteristics'),
       showContaminationBelowLifeblood: game.settings.get('twodsix', 'showContaminationBelowLifeblood'),
       showLifebloodStamina: game.settings.get("twodsix", "showLifebloodStamina"),
-      showHeroPoints: game.settings.get("twodsix", "showHeroPoints")
+      showHeroPoints: game.settings.get("twodsix", "showHeroPoints"),
+      showIcons: game.settings.get("twodsix", "showIcons")
     };
     data.config = TWODSIX;
 
@@ -76,7 +76,7 @@ export class TwodsixActorSheet extends AbstractTwodsixActorSheet {
     html.find('.rollable').on('click', this._onRollWrapper(this._onSkillRoll));
     html.find('.rollable-characteristic').on('click', this._onRollWrapper(this._onRollChar));
 
-    html.find('.roll-damage').on('click', this._onRollDamage.bind(this));
+    html.find('.roll-damage').on('click', onRollDamage.bind(this));
 
     html.find('#joat-skill-input').on('input', this._updateJoatSkill.bind(this));
     html.find('#joat-skill-input').on('blur', this._onJoatSkillBlur.bind(this));
@@ -186,24 +186,6 @@ export class TwodsixActorSheet extends AbstractTwodsixActorSheet {
     const fullCharLabel = getKeyByValue(TWODSIX.CHARACTERISTICS, shortChar);
     const displayShortChar = (<TwodsixActor>this.actor).data.data["characteristics"][fullCharLabel].displayShortLabel;
     await (<TwodsixActor>this.actor).characteristicRoll({ "characteristic": shortChar, "displayLabel": displayShortChar }, showThrowDiag);
-  }
-
-  /**
-   * Handle clickable damage rolls.
-   * @param {Event} event   The originating click event
-   * @private
-   */
-  private async _onRollDamage(event): Promise<void> {
-    event.preventDefault();
-    event.stopPropagation();
-    const itemId = $(event.currentTarget).parents('.item').data('item-id');
-    const item = this.actor.items.get(itemId) as TwodsixItem;
-
-    const element = $(event.currentTarget);
-    const bonusDamageFormula = String(element.data('bonus-damage') || 0);
-
-    await item.rollDamage((<DICE_ROLL_MODES>game.settings.get('core', 'rollMode')), bonusDamageFormula);
-
   }
 
   private static untrainedToJoat(skillValue: number): number {